Add in mutations for full forgotten password flow (#16)
* Add the forgotPassword migration.

This mutation always returns true to avoid information leakage (which
accounts exist that someone could then attempt to brute force passwords
on). Read more
https://www.troyhunt.com/everything-you-ever-wanted-to-know/ (a security
researcher and creator of HaveIBeenPwned) lists a number of things to
avoid, and username enumeration is one of them.

To make the user flow nice when they enter an email the have access to
we send an email saying "no user found" rather than presenting this to
the UI.

Similarly if they ask for a reset on an OAuth user we can't do that, so
tell them (again via email).

* Make resetToken unique on LocalCredential

We need to be able to find a single user from a password reset token, so
this should be unique so that we can pass it in the `query
localCredential`

Stop password being unique. Even though it very unlikely to clash
(because each password is stored as a salted bcrypt hash) it still isn't
an error if we happen to get two password hashes the same.

* Add resetPassword mutation to complete the forgotten password workflow

The UI of this could be improved in the future by possibly validating
the token _before_ presenting the user with a password reset form, and
the security here could be improved by:

- Time limiting the reset tokens (to an hour, say)
- Validating rules about the password on the server too (min length, not
  in top 1000 common most password for instance)