Email confirmation for users before they can sign in (#23)
* De-dup test code in create-token mutation a little bit

Each of the test cases had the same boiler plate - we can pull it out to
some variables in the shared block to reduce duplication

* Don't allow users to sign in if their account is not active

i.e. if they are banned, inactive or not-yet-verified.

* Automatically clear mocks between tests

This setting means that the number of calls/arguments mocks are called
with are reset for us between each test, but that the mock
implementation is kept (while it is in scope). This saves us having to
put `mockClear()` calls in our tests

* Send users a verification email before they can sign in.

To do that we need to:

- send confirmation emails
- create users in the pending state (except for OAuth users, which should be active straight
  away)
- not return a "token" in the AuthUser response if the user can't log in
- add mutations for confirmEmail and resendConfirmation.