Hacky-half-finished work against cascade-permissions branch
The idea behind this is to change the callers of
hasPermission/checkPermission back to pass entityType+id and instead
fetch all the needed info to compute the cascade when. This means that
the resolvers don't have to deal with fetching the object type _just_ to
pass it to hasPermission - instead the ctx.user.roleBindings has enough
(nested) info to work out what it needs.

:warning: This will almost certainl not pass tests or work as it is.