Fix ServiceAccount bugs to do with being able to edit the roles (#105)
* Validate the role type matches the entit type for SvcAccts

If they don't match then the all sorts of things will go wrong

* Update the role of the ServiceAccount when asked to

We were looking at `data.role`, but we didn't pick "role" out of the
payload, so this was always false. (And since data goes to the
updateServiceAccount mutation we don't pick it out here either, but look
at the local variable)

* Correctly set entityType field on ServiceAccount now roles can be set

The type would be null if the role was something other than Editor.
Since this is a required field we also change the schema type to reflect
this.