Change how we define our roles to make it easier to tweak permissions (#170)
The "config" module we use supports loading config files from multiple
sources, but the behavior with arrays is that you can't merge them, so
if you want to change any of the permissions at all you have to
duplicate the entire block of permissions to make a single change. This
makes upgrades error prone (and puts the onus on the admins of the
platform to merge changes correctly).

So instead this changes the layout form list-of-lists to dict-of-dicts,
as the config layering means we can tweak _exactly_ the permissions we
need, so for instance if you didn't want to allow Editors to push images
you could set:

    roles:
      DEPLOYMENT_EDITOR:
        permissions:
          deployment.images.push: false

and then _just_ that one permission is changed. This is easier to
maintain across upgrades as we add more permissions, and clearer to
understand what the intent is and what is actually changed.

In order to support anyone who currently has a custom role config we
upgrade the old config in memory before using it.