Ensure we run prisma deploy before using client library (#174)
Now that we are running  earlier, the prisma client library is updated by the time we run this migration and query for deployments (and create docker images). This means the client library and schema in the db are out of sync, which causes and issue when migrating from a previous version. This makes sure we run  before attempting to use the client library.